get the processes plotted
